jobpipe/hunt/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
requests>=2.32.0
python-dotenv>=1.0.1
supabase>=2.8.0
